+++ src/main/java/com/learnwebservices/LearnWebservicesApp.java
@@ -7,12 +7,14 @@ import com.learnwebservices.services.tempconverter.TempConverterEndpoint;
 import org.apache.cxf.Bus;
 import org.apache.cxf.ext.logging.LoggingFeature;
 import org.apache.cxf.jaxws.EndpointImpl;
+import org.apache.cxf.transport.servlet.CXFServlet;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.event.ApplicationStartedEvent;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
+import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.context.ApplicationListener;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -23,6 +25,7 @@ import org.springframework.web.filter.CorsFilter;
 
 import jakarta.xml.ws.Endpoint;
 import java.util.List;
+import java.util.Map;
 
 
 @SpringBootApplication
@@ -76,6 +79,13 @@ public class LearnWebservicesApp {
         FilterRegistrationBean<CorsFilter> bean = new FilterRegistrationBean<>(new CorsFilter(source));
         bean.setOrder(0);
         return bean;
+    }
+
+    @Bean
+    public ServletRegistrationBean<CXFServlet> cxfServlet() {
+        ServletRegistrationBean<CXFServlet> servletRegistrationBean = new ServletRegistrationBean<>(new CXFServlet(), "/services/*");
+        servletRegistrationBean.setInitParameters(Map.of("use-x-forwarded-headers", "true"));
+        return servletRegistrationBean;
     }
 
 }